[ADD] Estate: Capitulo 14 y 15
Se agreraron los ejercicios de los capitulos 14 y 15
del Getting Started de Odoo, junto con el uso de la
herramineta pre-commit-vauxoo
