--- __manifest__.py
@@ -1,18 +1,0 @@
-{
-    "name": "Real Estate",
-    "depends": [
-      "base",
-    ],
-    "data": [
-      "security/ir.model.access.csv",
-      "views/estate_property_views.xml",
-      "views/estate_property_type_views.xml",
-      "views/estate_property_tag_views.xml",
-      "views/estate_property_offer_views.xml",
-      "views/res_users_view.xml",
-      "views/estate_menus.xml"
-    ],
-    "installable": True,
-    "application": True,
-    "auto_install": False
-}
