+++ estate/models/res_users.py
@@ -1,0 +1,8 @@
+from odoo import fields, models
+
+
+class ResUsers(models.Model):
+    _inherit = "res.users"
+
+    property_ids = fields.One2many("estate.property", "user_id")
+    domain = ["|", ("state", "=", "new"), ("state", "=", "offer received")]
