+++ estate/views/estate_property_tag_views.xml
@@ -1,4 +1,4 @@
-<?xml version="1.0"?>
+<?xml version="1.0" ?>
 <odoo>
 
   <record id="estate_property_tag_action" model="ir.actions.act_window">
@@ -14,8 +14,8 @@
       <form string="Test">
         <sheet>
           <div>
-            <label for="name"/>
-            <field class="mb16" name="name"/>
+            <label for="name" />
+            <field class="mb16" name="name" />
           </div>
         </sheet>
       </form>
@@ -27,9 +27,9 @@
     <field name="model">estate.property.tag</field>
     <field name="arch" type="xml">
       <tree string="Channel" editable="bottom">
-        <field name="name"/>
+        <field name="name" />
       </tree>
     </field>
   </record>
-  
+
 </odoo>
