+++ estate/views/estate_property_type_views.xml
@@ -1,4 +1,4 @@
-<?xml version="1.0"?>
+<?xml version="1.0" ?>
 <odoo>
 
   <record id="estate_property_type_action" model="ir.actions.act_window">
@@ -12,8 +12,8 @@
     <field name="model">estate.property.type</field>
     <field name="arch" type="xml">
       <tree>
-        <field name="sequence" widget="handle"/>
-        <field name="name"/>
+        <field name="sequence" widget="handle" />
+        <field name="name" />
       </tree>
     </field>
   </record>
