+++ estate/views/res_users_view.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+  <data>
+
+    <record id="res_users_view_form" model="ir.ui.view">
+      <field name="name">res.users.view.form.inherit.estate</field>
+      <field name="model">res.users</field>
+      <field name="inherit_id" ref="base.view_users_form" />
+      <field name="arch" type="xml">
+        <notebook>
+          <page string="Properties">
+            <field name="property_ids">
+              <tree
+                                decoration-success="state in ('offer_received', 'offer_accepted')"
+                                decoration-muted="state == 'sold'"
+                                decoration-bf="state == 'offer_accepted'"
+                            >
+                <field name="name" string="Title" />
+                <field name="postcode" />
+                <field name="bedrooms" />
+                <field name="state" />
+                <field name="living_area" string="Living Area (sqm)" />
+                <field name="expected_price" />
+                <field name="selling_price" />
+                <field name="date_availability" string="Available From" optional="hide" />
+              </tree>
+            </field>
+          </page>
+        </notebook>
+      </field>
+    </record>
+
+  </data>
+</odoo>
