+++ estate_account/__init__.py
@@ -1,0 +1,1 @@
+from . import models
