+++ estate_account/__manifest__.py
@@ -1,0 +1,13 @@
+{
+    "name": "Estate Account",
+    "depends": [
+      "base",
+      "estate",
+      "account"
+    ],
+    "data": [
+    ],
+    "installable": True,
+    "application": True,
+    "auto_install": False
+}
