--- models/estate_property_type.py
@@ -1,27 +1,0 @@
-from odoo import fields, models, api
-
-class EstatePropertyType(models.Model):
-  _name = "estate.property.type"
-  _description = "Property Type Model"
-  _order = "name"
-
-  name = fields.Char(required = True)
-  property_ids = fields.One2many("estate.property", "property_type_id")
-  sequence = fields.Integer('Sequence')
-  offer_ids = fields.One2many("estate.property.offer", "property_type_id")
-  offer_count = fields.Integer(compute = "_compute_offer_count")
-  
-  _sql_constraints = [
-    ("unique_type_name",
-    "UNIQUE(name)",
-    "Type name alredy exists")
-  ]
-
-  #private methods
-  @api.depends("offer_ids")
-  def _compute_offer_count(self):
-    for record in self:
-      if record.offer_ids:
-        record.offer_count = len(record.offer_ids)
-      else:
-        record.offer_count = 0
