--- views/estate_menus.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0"?>
-<odoo>
-
-  <menuitem id="estate_property_root" name="Real Estate">
-    <menuitem id="estate_first_level_menu" name="Advertisements">
-      <menuitem action="estate_property_action" id="estate_property_menu_action"/>
-    </menuitem>
-
-    <menuitem id="estate_property_first_level_menu" name="Settings">
-      <menuitem action="estate_property_type_action" id="estate_property_type_menu_action"/>
-      <menuitem action="estate_property_tag_action" id="estate_property_tag_menu_action"/>
-    </menuitem>
-  </menuitem>
-  
-</odoo>
