--- views/estate_property_views.xml
@@ -1,192 +1,0 @@
-<?xml version="1.0"?>
-<odoo>
-
-  <record id="estate_property_action" model="ir.actions.act_window">
-    <field name="name">Properties</field>
-    <field name="res_model">estate.property</field>
-    <field name="view_mode">tree,form</field>
-    <field name="context">{'search_default_date_availability': True, 'search_default_current': True}</field>
-  </record>
-
-  <record id="estate_property_view_tree" model="ir.ui.view">
-    <field name="name">estate.property.tree</field>
-    <field name="model">estate.property</field>
-    <field name="arch" type="xml">
-      <tree decoration-success="state in ('offer_received', 'offer_accepted')" decoration-muted="state == 'sold'" decoration-bf="state == 'offer_accepted'">
-        <field name="name" string="Title"/>
-        <field name="postcode"/>
-        <field name="bedrooms"/>
-        <field name="state"/>
-        <field name="living_area" string="Living Area (sqm)"/>
-        <field name="expected_price"/>
-        <field name="selling_price"/>
-        <field name="date_availability" string="Available From" optional="hide"/>
-      </tree>
-    </field>
-  </record>
-
-  <record id="estate_property_view_form" model="ir.ui.view">
-    <field name="name">estate.property.form</field>
-    <field name="model">estate.property</field>
-    <field name="arch" type="xml">
-      <form string="Test">
-        <header>
-          <button name="action_sold_property" string="Sold" type="object" states="new,offer received"/>
-          <button name="action_sold_property" string="Sold" type="object" states="offer_accepted" class="oe_highlight"/>
-          <button name="action_cancel_property" string="Cancel" type="object" states="new,offer_received,offer_accepted"/>
-          <field name="state" widget="statusbar" statusbar_visible="new,offer_received,offer_accepted,sold"/>
-        </header>
-        <sheet>
-          <div>
-            <h1 class="mb32">
-              <field class="mb16" name="name"/>
-            </h1>
-          </div>
-          <div class="col-md-6">
-            <group>
-              <field name="tag_ids" widget="many2many_tags" options="{'color_field': 'color'}"/>
-            </group>
-          </div>
-          <div style="display: flex;">
-            <div>
-              <div>
-                <label for="property_type_id"/>
-                <field class="mb16" name="property_type_id"  options="{'no_create': True}"/>
-              </div>
-              <div>
-                <label for="postcode"/>
-                <field class="mb16" name="postcode"/>
-              </div>
-              <div>
-                <label for="date_availability"/>
-                <field class="mb16" name="date_availability"/>
-              </div>
-            </div>
-            <div>
-              <div>
-                <label for="expected_price"/>
-                <field class="mb16" name="expected_price"/>
-              </div>
-              <div>
-                <label for="best_price"/>
-                <field class="mb16" name="best_price"/>
-              </div>
-              <div>
-                <label for="selling_price"/>
-                <field class="mb16" name="selling_price"/>
-              </div>
-            </div>
-          </div>
-          <notebook>
-            <page string="Description">
-              <div style="display: flex;">
-                <div style="width: 120px">
-                  <label for="description"/>
-                </div>
-                <div>
-                  <field name="description" style="width: 30vw;"/>
-                </div>
-              </div>
-              <div style="display: flex;">
-                <div style="width: 120px">
-                  <label for="living_area" string="Living Area (sqm)"/>
-                </div>
-                <div>
-                  <field class="mb16" name="living_area"/>
-                </div>
-              </div>
-              <div style="display: flex;">
-                <div style="width: 120px">
-                  <label for="facades"/>
-                </div>
-                <div>
-                  <field class="mb16" name="facades"/>
-                </div>
-              </div>
-              <div style="display: flex;">
-                <div style="width: 120px">
-                  <label for="garage"/>
-                </div>
-                <div style="width: 120px">
-                  <field class="mb16" name="garage"/>
-                </div>
-              </div>
-              <div style="display: flex;">
-                <div style="width: 120px">
-                  <label for="garden" />
-                </div>
-                <div>
-                  <field class="mb16" name="garden"/>
-                </div>
-              </div>
-              <div style="display: flex;" attrs="{'invisible': [('garden', '=', False)]}">
-                <div style="width: 120px">
-                  <label for="garden_area" string="Garden Area (sqm)"/>
-                </div>
-                <div>
-                  <field class="mb16" name="garden_area"/>
-                </div>
-              </div>
-              <div style="display: flex;" attrs="{'invisible': [('garden', '=', False)]}">
-                <div style="width: 120px">
-                  <label for="garden_orientation"/>
-                </div>
-                <div>
-                  <field class="mb16" name="garden_orientation"/>
-                </div>
-              </div>
-              <div style="display: flex;">
-                <div style="width: 120px">
-                  <label for="total_area"/>
-                </div>
-                <div>
-                  <field class="mb16" name="total_area"/>
-                </div>
-              </div>
-            </page>
-            <page string="Offers">
-              <div style="display: flex;">
-                <div>
-                  <field name="offer_ids" attrs="{'readonly':['|', ('state', '=','sold'), ('state', '=','canceled')]}" editable="top"/>
-                </div>
-              </div>
-            </page>
-            <page string="Other Info">
-              <div style="display: flex;">
-                <div style="width: 110px">
-                  <label for="user_id"/>
-                </div>
-                <div>
-                  <field name="user_id" style="width: 30vw;"/>
-                </div>
-                <div style="width: 110px">
-                  <label for="partner_id"/>
-                </div>
-                <div>
-                  <field name="partner_id" style="width: 30vw;"/>
-                </div>
-              </div>
-            </page>
-          </notebook>
-        </sheet>
-      </form>
-    </field>
-  </record>
-
-  <record id="estate_property_view_search" model="ir.ui.view">
-    <field name="name">estate.property.search</field>
-    <field name="model">estate.property</field>
-    <field name="arch" type="xml">
-      <search string="Estate Property">
-        <field name="name" string="Title"/>
-        <field name="postcode"/>
-        <field name="expected_price"/>
-        <field name="bedrooms"/>
-        <field name="living_area" string="Living Area (sqm)" filter_domain="[('living_area', 'ilike', self)]"/>
-        <field name="facades"/>
-        <filter domain="['|',('state', '=', 'new'),('state', '=', 'offer received')]" name="date_availability" string="Available" />
-      </search>
-    </field>
-  </record>
-  
-</odoo>
